+++ src/ca/akennedy/gents/weightmatrix/WeightFeaturesRelationSupervised.java
@@ -150,8 +150,8 @@ public class WeightFeaturesRelationSupervised extends WeightFeaturesContextSuper
 						double value = 1;
 						boolean isGood = true;
 						value = MatrixWeighter.getAssociation(truePos, falsePos, falseNeg, trueNeg, type);
-						//if(falseNeg > 0 && (Double.isNaN(value) || Double.isInfinite(value))){
-							LOGGER.severe("Error at feature: " + relation + "\tnumber: "+featureNumber +"\n" +(long)truePos + " " + (long)falsePos + "\n" + (long)falseNeg + " " + (long)trueNeg + "\n" + value);
+						//if(Double.isNaN(value) || Double.isInfinite(value) || trueNeg < 0){
+							LOGGER.severe("Error at feature: " + relation + "\tnumber: "+featureNumber +"\n" +truePos + " " + falsePos + "\n" + falseNeg + " " + trueNeg + "\n" + value);
 						//}
 						if(truePos + falsePos + trueNeg + falseNeg == 0){
 							isGood = false;
